new DocsTable component
