+++ packages/styleguide/src/components/DocsTable/index.js
@@ -1,0 +1,1 @@
+export { default } from './DocsTable';
