+++ packages/styleguide/src/components/Table/index.js
@@ -1,1 +1,2 @@
 export { default } from './Table';
+export { propTypes as TablePropTypes } from './Table';
